[ADMIN] Move files on user into users folder
