+++ admin/app/components/users/user-detail-personal-information/schooling-registration-information.hbs
