+++ admin/app/styles/app.scss
@@ -57,7 +57,7 @@
 @import 'components/publish-session-button';
 @import 'components/sessions/jury-comment';
 @import 'components/sessions/list-items.scss';
-@import 'components/user-detail-personal-information-section';
+@import 'components/users/user-detail-personal-information-section';
 @import 'components/target-profiles/badges';
 @import 'components/target-profiles/target-profile';
 @import 'components/target-profiles/badge-form';
