+++ admin/tests/unit/components/users/user-detail-personal-information/user-overview_test.js
@@ -1,15 +1,15 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
-import createGlimmerComponent from '../../../helpers/create-glimmer-component';
+import createGlimmerComponent from '../../../../helpers/create-glimmer-component';
 import ENV from 'pix-admin/config/environment';
 
-module('Unit | Component | user-detail-personal-information', function (hooks) {
+module('Unit | Component | users | user-detail-personal-information', function (hooks) {
   setupTest(hooks);
 
   module('#externalURL', function () {
     test('it should generate dashboard URL based on environment and object', async function (assert) {
       // given
-      const component = createGlimmerComponent('component:user-detail-personal-information/user-overview');
+      const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
 
       const args = {
         user: {
@@ -34,7 +34,7 @@ module('Unit | Component | user-detail-personal-information', function (hooks) {
     module('when user already has an email', function () {
       test('it should allow email modification', async function (assert) {
         // given
-        const component = createGlimmerComponent('component:user-detail-personal-information/user-overview');
+        const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
         const user = { email: 'lisa@example.net', firstName: 'Lisa', lastName: 'Dupont' };
         component.args.user = user;
 
@@ -46,7 +46,7 @@ module('Unit | Component | user-detail-personal-information', function (hooks) {
     module('when user has an username', function () {
       test('it should also allow email modification', async function (assert) {
         // given
-        const component = createGlimmerComponent('component:user-detail-personal-information/user-overview');
+        const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
         const user = { username: 'lisa.dupont', firstName: 'Lisa', lastName: 'Dupont' };
         component.args.user = user;
 
@@ -58,7 +58,7 @@ module('Unit | Component | user-detail-personal-information', function (hooks) {
     module('when user has neither a username nor an email', function () {
       test('it should not allow email modification', async function (assert) {
         // given
-        const component = createGlimmerComponent('component:user-detail-personal-information/user-overview');
+        const component = createGlimmerComponent('component:users/user-detail-personal-information/user-overview');
         const user = { firstName: 'Lisa', lastName: 'Dupont' };
         component.args.user = user;
